Split SetHeader into desktop and classic versions, add loading skeletons where appropriate (#268)
